+++ core/primitives/src/challenge.rs
@@ -1,49 +1,14 @@
 use crate::hash::CryptoHash;
 use crate::merkle::MerklePath;
 use crate::sharding::{EncodedShardChunk, ShardChunk, ShardChunkHeader};
+use crate::state::PartialState;
 use crate::types::AccountId;
 use crate::validator_signer::ValidatorSigner;
 use borsh::{BorshDeserialize, BorshSerialize};
 use near_crypto::Signature;
 use near_primitives_core::types::BlockHeight;
 use near_schema_checker_lib::ProtocolSchema;
-use std::fmt::{Debug, Formatter};
-
-/// Serialized TrieNodeWithSize or state value.
-pub type TrieValue = std::sync::Arc<[u8]>;
-
-#[derive(BorshSerialize, BorshDeserialize, Clone, Eq, PartialEq, ProtocolSchema)]
-/// TODO (#8984): consider supporting format containing trie values only for
-/// state part boundaries and storing state items for state part range.
-pub enum PartialState {
-    /// State represented by the set of unique trie values (`RawTrieNodeWithSize`s and state values).
-    TrieValues(Vec<TrieValue>),
-}
-
-impl Default for PartialState {
-    fn default() -> Self {
-        PartialState::TrieValues(vec![])
-    }
-}
-
-// When debug-printing, don't dump the entire partial state; that is very unlikely to be useful,
-// and wastes a lot of screen space.
-impl Debug for PartialState {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            PartialState::TrieValues(values) => {
-                f.write_str(&format!("{} trie values", values.len()))
-            }
-        }
-    }
-}
-
-impl PartialState {
-    pub fn len(&self) -> usize {
-        let Self::TrieValues(values) = self;
-        values.len()
-    }
-}
+use std::fmt::Debug;
 
 /// Double signed block.
 #[derive(BorshSerialize, BorshDeserialize, PartialEq, Eq, Clone, Debug, ProtocolSchema)]
