+++ core/store/src/trie/trie_storage.rs
@@ -7,9 +7,9 @@ use lru::LruCache;
 use near_o11y::log_assert;
 use near_o11y::metrics::prometheus;
 use near_o11y::metrics::prometheus::core::{GenericCounter, GenericGauge};
-use near_primitives::challenge::PartialState;
 use near_primitives::hash::CryptoHash;
 use near_primitives::shard_layout::ShardUId;
+use near_primitives::state::PartialState;
 use near_primitives::types::ShardId;
 use std::collections::{HashMap, HashSet, VecDeque};
 use std::num::NonZeroUsize;
