+++ core/store/src/trie/trie_tests.rs
@@ -3,10 +3,10 @@ use crate::trie::trie_storage::{TrieMemoryPartialStorage, TrieStorage};
 use crate::trie::TrieNodesCount;
 use crate::{PartialStorage, Trie, TrieUpdate};
 use assert_matches::assert_matches;
-use near_primitives::challenge::PartialState;
 use near_primitives::errors::{MissingTrieValueContext, StorageError};
 use near_primitives::hash::{hash, CryptoHash};
 use near_primitives::shard_layout::{ShardLayout, ShardUId};
+use near_primitives::state::PartialState;
 use rand::seq::SliceRandom;
 use rand::Rng;
 use std::collections::{HashMap, HashSet};
